aider: Removed 0 from the bottom right tile.
# Aider chat conversation:

USER: can you remove the 0 from the bottom right tile?

ASSISTANT: Sure, we can modify the `script.js` file to not display the number 0 on the puzzle piece.

Here are the changes:

```javascript
script.js
<<<<<<< HEAD
piece.textContent = puzzle[i][j];
=======
piece.textContent = puzzle[i][j] === 0 ? '' : puzzle[i][j];
>>>>>>> updated
```

This change will set the text content of the puzzle piece to an empty string if the number is 0, otherwise it will use the number.