aider: Fixed issue with double clicking a piece causing it to disappear.
# Aider chat conversation:

USER: I double clicked 15 and it disappeared ... that is not the intended behaviour

ASSISTANT: I see. The issue seems to be that the `mousedown` event is triggering the `movePiece` function even when the piece is not adjacent to the empty tile. This is because the `mouseup` event triggers the `movePiece` function, and the `mousedown` event sets the `draggedPiece` to the piece that was clicked. If you double click a piece, the `mousedown` event will trigger twice, setting the `draggedPiece` to the piece that was clicked, and then the `mouseup` event will trigger the `movePiece` function, causing the piece to disappear.

To fix this, we need to modify the `mousedown` event listener to only set the `draggedPiece` if the piece is adjacent to the empty tile. We can use the `movePiece` function to check if a piece is adjacent to the empty tile, but we need to modify it to return a boolean value indicating whether the piece was moved or not.

Here are the changes we need to make to the `script.js` file:

1. Modify the `movePiece` function to return a boolean value indicating whether the piece was moved or not.
2. Modify the `mousedown` event listener to only set the `draggedPiece` if the piece is adjacent to the empty tile.

Here are the *edit blocks* for these changes:

```javascript
script.js
<<<<<<< HEAD
function movePiece(piece) {
    var pieceNumber = parseInt(piece.textContent);
    var piecePosition;
    var emptyTilePosition = getEmptyTilePosition();

    for (var i = 0; i < 4; i++) {
        for (var j = 0; j < 4; j++) {
            if (puzzle[i][j] === pieceNumber) {
                piecePosition = { i: i, j: j };
                break;
            }
        }
    }

    if (Math.abs(piecePosition.i - emptyTilePosition.i) + Math.abs(piecePosition.j - emptyTilePosition.j) === 1) {
        puzzle[piecePosition.i][piecePosition.j] = 0;
        puzzle[emptyTilePosition.i][emptyTilePosition.j] = pieceNumber;
        piece.style.left = (emptyTilePosition.j * 100) + 'px';
        piece.style.top = (emptyTilePosition.i * 100) + 'px';
    }
}
=======
function movePiece(piece) {
    var pieceNumber = parseInt(piece.textContent);
    var piecePosition;
    var emptyTilePosition = getEmptyTilePosition();

    for (var i = 0; i < 4; i++) {
        for (var j = 0; j < 4; j++) {
            if (puzzle[i][j] === pieceNumber) {
                piecePosition = { i: i, j: j };
                break;
            }
        }
    }

    if (Math.abs(piecePosition.i - emptyTilePosition.i) + Math.abs(piecePosition.j - emptyTilePosition.j) === 1) {
        puzzle[piecePosition.i][piecePosition.j] = 0;
        puzzle[emptyTilePosition.i][emptyTilePosition.j] = pieceNumber;
        piece.style.left = (emptyTilePosition.j * 100) + 'px';
        piece.style.top = (emptyTilePosition.i * 100) + 'px';
        return true;
    }

    return false;
}
>>>>>>> updated
```

```javascript
script.js
<<<<<<< HEAD
piece.addEventListener('mousedown', function(e) {
    dragging = true;
    draggedPiece = this;
    e.preventDefault();
});
=======
piece.addEventListener('mousedown', function(e) {
    if (movePiece(this)) {
        dragging = true;
        draggedPiece = this;
        e.preventDefault();
    }
});
>>>>>>> updated
```