aider: Changed mousedown event listener to call movePiece function when a piece is clicked.
# Aider chat conversation:

USER: 15 is next to blank. if I click 15, 15 and Blank should change place. can you do that?

ASSISTANT: Sure, I can help with that. We need to modify the `mousedown` event listener to call the `movePiece` function when a piece is clicked. If the piece is adjacent to the empty tile, they will swap places.

Here is the *edit block* for this change:

```javascript
script.js
<<<<<<< HEAD
piece.addEventListener('mousedown', function(e) {
    if (movePiece(this)) {
        dragging = true;
        draggedPiece = this;
        e.preventDefault();
    }
});
=======
piece.addEventListener('mousedown', function(e) {
    movePiece(this);
    e.preventDefault();
});
>>>>>>> updated
```

Please make this change to the `script.js` file and let me know if the issue is resolved.