aider: The text content of the blank tile is now hidden by setting its color to be the same as the background color.
# Aider chat conversation:

USER: I have """            if (puzzle[i][j] === 0) {
                piece.classList.add('blank-tile');
            }
""" but the tile 0 shows the 0 and not the blank

ASSISTANT: The issue is that the text content of the tile is still being set to the number 0, even though the `blank-tile` class is being added. We need to modify the `script.js` file to not set the text content for the blank tile.

Here is the change:

```javascript
script.js
<<<<<<< HEAD
            if (puzzle[i][j] === 0) {
                piece.classList.add('blank-tile');
            }
            piece.textContent = puzzle[i][j];
=======
            if (puzzle[i][j] === 0) {
                piece.classList.add('blank-tile');
            } else {
                piece.textContent = puzzle[i][j];
            }
>>>>>>> updated
```

This change will ensure that the text content for
^C KeyboardInterrupt

USER: no, i need the content to be 0 for later, I just dont want to show it in the browser

ASSISTANT: I see, you want to keep the value of the tile as 0 in the JavaScript code, but you don't want to display it in the browser. In that case, you can use CSS to hide the text of the blank tile.

We need to modify the `styles.css` file to add a new rule that sets the color of the text in the blank tile to be the same as the background color, effectively hiding it.

Here is the change:

```css
styles.css
<<<<<<< HEAD
.blank-tile {
    background-color: #fff;
}
=======
.blank-tile {
    background-color: #fff;
    color: #fff;
}
>>>>>>> updated
```

This change will make the text of the blank tile the same color as the background, effectively hiding it.