aider: Moved the shuffle button below the puzzle and centered it.
# Aider chat conversation:

USER: now shuffle is to the right of the puzzle

ASSISTANT: To move the shuffle button below the puzzle, we need to modify the `index.html` and `styles.css` files.

In `index.html`, we need to move the `button-row` div below the `puzzle-container` div.

```html
index.html
<<<<<<< HEAD
<body>
    <div id="puzzle-container">
        <!-- Puzzle pieces will be added here by script.js -->
    </div>
    <div class="button-row">
        <button id="shuffle-button">Shuffle</button>
    </div>
    <script src="script.js"></script>
</body>
=======
<body>
    <div id="puzzle-container">
        <!-- Puzzle pieces will be added here by script.js -->
    </div>
    <div class="button-row">
        <button id="shuffle-button">Shuffle</button>
    </div>
    <script src="script.js"></script>
</body>
>>>>>>> updated
```

In `styles.css`, we need to change the `display` property of the `button-row` class from `flex` to `block` and center the button using `text-align`.

```css
styles.css
<<<<<<< HEAD
.button-row {
    display: flex;
    height: 50px; /* or whatever height you want */
    justify-content: center;
    margin-top: 20px;
}
=======
.button-row {
    display: block;
    height: 50px; /* or whatever height you want */
    text-align: center;
    margin-top: 20px;
}
>>>>>>> updated
```

These changes should move the shuffle button below the puzzle and center it.