aider: Added a new button "14" which swaps the positions of "14" and "15" in the puzzle array when clicked.
# Aider chat conversation:

USER: can you make a new button "14" which creates the standard starting position with "14" and "15" swapped?

ASSISTANT: Sure, we can add a new button "14" to the HTML structure and add an event listener to it in the JavaScript file.

First, let's add the new button to the `index.html` file.

```html
index.html
<<<<<<< HEAD
<body>
    <div id="puzzle-container">
        <!-- Puzzle pieces will be added here by script.js -->
    </div>
    <div class="button-row">
        <button id="shuffle-button">Shuffle</button>
    </div>
    <script src="script.js"></script>
</body>
=======
<body>
    <div id="puzzle-container">
        <!-- Puzzle pieces will be added here by script.js -->
    </div>
    <div class="button-row">
        <button id="shuffle-button">Shuffle</button>
        <button id="swap-button">14</button> <!-- New button -->
    </div>
    <script src="script.js"></script>
</body>
>>>>>>> updated
```

Next, let's add an event listener to the new button in the `script.js` file. This event listener will swap the positions of "14" and "15" in the puzzle array.

```javascript
script.js
<<<<<<< HEAD
document.getElementById('shuffle-button').addEventListener('click', shufflePuzzle);
=======
document.getElementById('shuffle-button').addEventListener('click', shufflePuzzle);
document.getElementById('swap-button').addEventListener('click', function() {
    // Swap "14" and "15" in the puzzle array
    var temp = puzzle[3][2];
    puzzle[3][2] = puzzle[3][3];
    puzzle[3][3] = temp;
    updatePuzzlePieces();
});
>>>>>>> updated
```

These changes will add a new button "14" to the HTML structure and make it swap the positions of "14" and "15" in the puzzle when clicked.