aider: Swapped "14" and "15" and moved "0" tile to the bottom right.
# Aider chat conversation:

USER: almost, but I want the 0-tile in the bottom right, 14 to the left, and 15 to the left

ASSISTANT: To achieve this, we need to modify the `script.js` file. We will change the initial puzzle array to have "14" and "15" swapped and the "0" tile in the bottom right.

Here is the necessary change:

```javascript
script.js
<<<<<<< HEAD
var puzzle = [
    [1, 2, 3, 4],
    [5, 6, 7, 8],
    [9, 10, 11, 12],
    [13, 14, 15, 0]
];
=======
var puzzle = [
    [1, 2, 3, 4],
    [5, 6, 7, 8],
    [9, 10, 11, 12],
    [13, 15, 14, 0]
];
>>>>>>> updated
```

This change will set the initial state of the puzzle to have "14" and "15" swapped and the "0" tile in the bottom right.