aider: Add new function `getParity(pieces)` to calculate the parity based on the order of the pieces.
# Aider chat conversation:

ASSISTANT: Sure, we can modify the function to take the actual pieces as input. The pieces would be an array of the numbers on the tiles in their current order. Here is the updated function:

```javascript
script.js
<<<<<<< HEAD
function getParity(puzzle) {
    var inversions = 0;
    var flatPuzzle = puzzle.flat();
    for (var i = 0; i < flatPuzzle.length - 1; i++) {
        for (var j = i + 1; j < flatPuzzle.length; j++) {
            if (flatPuzzle[i] > flatPuzzle[j] && flatPuzzle[i] != 0 && flatPuzzle[j] != 0) {
                inversions++;
            }
        }
    }
    return inversions % 2;
}
=======
function getParity(pieces) {
    var inversions = 0;
    for (var i = 0; i < pieces.length - 1; i++) {
        for (var j = i + 1; j < pieces.length; j++) {
            if (pieces[i] > pieces[j] && pieces[i] != 0 && pieces[j] != 0) {
                inversions++;
            }
        }
    }
    return inversions % 2;
}
>>>>>>> updated
```

This function now takes an array of pieces as input, and calculates the parity based on the order of these pieces.

USER: InvalidEditBlock: edit failed!

script.js does not contain the *exact sequence* of HEAD lines you specified.
Try again.
DO NOT skip blank lines, comments, docstrings, etc!
The HEAD block needs to be EXACTLY the same as the lines in script.js with nothing missing!

script.js does not contain these 12 exact lines in a row:
```
function getParity(puzzle) {
    var inversions = 0;
    var flatPuzzle = puzzle.flat();
    for (var i = 0; i < flatPuzzle.length - 1; i++) {
        for (var j = i + 1; j < flatPuzzle.length; j++) {
            if (flatPuzzle[i] > flatPuzzle[j] && flatPuzzle[i] != 0 && flatPuzzle[j] != 0) {
                inversions++;
            }
        }
    }
    return inversions % 2;
}
```

ASSISTANT: I apologize for the confusion. It seems there was a misunderstanding. The function `getParity(puzzle)` does not exist in the provided `script.js` file. Therefore, we can't modify it. Instead, we can add the new function `getParity(pieces)` to the `script.js` file. Here is the edit block for adding the new function:

```javascript
script.js
<<<<<<< HEAD

=======
function getParity(pieces) {
    var inversions = 0;
    for (var i = 0; i < pieces.length - 1; i++) {
        for (var j = i + 1; j < pieces.length; j++) {
            if (pieces[i] > pieces[j] && pieces[i] != 0 && pieces[j] != 0) {
                inversions++;
            }
        }
    }
    return inversions % 2;
}
>>>>>>> updated
```

This function takes an array of pieces as input, and calculates the parity based on the order of these pieces.