aider: Added default text and changed color of `#parity-display` element to red.
# Aider chat conversation:

USER: i cannot see the element

ASSISTANT: If you can't see the `#parity-display` element, it might be because it's empty or its color is blending with the background.

Let's add some default text to the `#parity-display` element in the HTML file and change its color in the CSS file to ensure it's visible.

Here are the changes:

**index.html**
```html
index.html
<<<<<<< HEAD
    <div id="parity-display"></div>
=======
    <div id="parity-display">Parity: </div>
>>>>>>> updated
```

**styles.css**
```css
styles.css
<<<<<<< HEAD
    #parity-display {
        position: absolute;
        top: 0;
        right: 0;
        font-size: 1.5em;
        color: #000;
    }
=======
    #parity-display {
        position: absolute;
        top: 0;
        right: 0;
        font-size: 1.5em;
        color: #ff0000; /* Change the color to red */
    }
>>>>>>> updated
```

These changes will add default text to the `#parity-display` element and change its color to red, making it more visible.